Make m_flags an int instead of a short, this is consistent with the type of the 'flags' argument m_getcl() was using anyway; m_extadd() needed to be changed to accept an int instead of a short for 'flags.' This makes things more consistent and also gives us more bits to use for m_flags in the future (we have almost run out).
Requested by: sam (Sam Leffler)

+++ sys/kern/subr_mbuf.c
@@ -1511,7 +1511,7 @@ m_clget(struct mbuf *mb, int how)
  */
 void
 m_extadd(struct mbuf *mb, caddr_t buf, u_int size,
-    void (*freef)(void *, void *), void *args, short flags, int type)
+    void (*freef)(void *, void *), void *args, int flags, int type)
 {
 
 	_mext_init_ref(mb, ((type != EXT_CLUSTER) ?
